pyproject.toml
[tool.poetry]
name = "Time track system"
version = "0.1.0"
description = "Time track system"
authors = ["qsethuk <alexfretka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
Django = "5.0.1"
cookiecutter = "^2.5.0"
django-jazzmin = "^2.6.0"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
sphinx = "^7.2.6"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.1.11"
djlint = "^1.34.1"
pylint = "^3.0.3"
coverage = "^7.4.0"
crispy-bootstrap5 = "^2023.10"
psycopg-binary = "^3.1.16"
django-allauth = "^0.60.0"
django-debug-toolbar = "^4.2.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
django-extensions = "^3.2.3"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
argon2-cffi = "^23.1.0"
jinja2 = "^3.1.2"
django-model-utils = "^4.3.1"
django-crispy-forms = "^2.1"
python-slugify = "^8.0.1"
pillow = "^10.2.0"
uvicorn = "^0.25.0"
werkzeug = "^3.0.1"
ipdb = "^0.13.13"
watchfiles = "^0.21.0"
pytest-sugar = "^0.9.7"
djangorestframework-stubs = "^3.14.5"
django-stubs = "^4.2.7"
factory-boy = "^3.3.0"
django-coverage-plugin = "^3.1.0"
pytest-django = "^4.7.0"
pylint-django = "^2.5.5"
setuptools = "^69.0.3"
djangoql = "^0.18.1"
django-easy-audit = { git = "https://github.com/soynatan/django-easy-audit.git" }
stubs = "^1.0.0"
django-filter = "^23.5"
colorama = "^0.4.6"
python-calamine = "^0.1.7"
django-sortedm2m = "^3.1.1"
line-profiler = "^4.1.2"
scalene = "^1.5.33.1"
psycopg2 = "^2.9.9"
django-bulk-update-or-create = "^0.3.0"
django-adminactions = "^2.3.0"
django-rosetta = "^0.10.0"
pyasyncore = "^1.0.3"
pytest-cov = "^4.1.0"
pytest = "^8.0.0"
faker = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==== Coverage ====
[tool.coverage.run]
include = ["src"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]


# ==== black ====
[tool.black]
line-length = 119
target-version = ['py312']


# ==== isort ====
[tool.isort]
profile = "black"
line_length = 119
known_first_party = [
    "src",
]
skip = [".venv/"]
skip_glob = ["**/migrations/*.py"]


# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"


# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
